+++ frontend/src/data/Templates.js
@@ -1,0 +1,35 @@
+const templates = [
+  {
+    title: 'Office',
+    description: 'Corporate and professional with a sleek modern look.',
+    image: '/images/office.png',
+  },
+  {
+    title: 'Studio',
+    description: 'A clean, minimal space perfect for designers and creatives.',
+    image: '/images/studio.png',
+  },
+  {
+    title: 'Gallery',
+    description: 'A curated portfolio display, like a personal museum.',
+    image: '/images/gallery.png',
+  },
+  {
+    title: 'Space',
+    description: 'A futuristic scene that showcases innovation and ambition.',
+    image: '/images/space.png',
+    href: '/space',
+  },
+  {
+    title: 'Forest',
+    description: 'Serene and natural — ideal for calm, grounded presentation.',
+    image: '/images/forest.png',
+  },
+  {
+    title: 'Cafe',
+    description: 'A cozy, warm atmosphere perfect for creative professionals.',
+    image: '/images/cafe.png',
+  },
+];
+
+export default templates;
